Generalized some mrna classes; improve MappingsFileRowValidator
