+++ src/main/groovy/org/transmartproject/batch/highdim/compute/DataPoint.groovy
@@ -1,0 +1,8 @@
+package org.transmartproject.batch.highdim.compute
+
+/**
+ * A data point that can be fed to the
+ */
+interface DataPoint {
+    Double getValue()
+}
