+++ src/main/groovy/org/transmartproject/batch/highdim/i2b2/ValidateHighDimensionalConceptsTasklet.groovy
@@ -1,4 +1,4 @@
-package org.transmartproject.batch.highdim.mrna.data.mapping
+package org.transmartproject.batch.highdim.i2b2
 
 import org.springframework.batch.core.StepContribution
 import org.springframework.batch.core.scope.context.ChunkContext
